finetuning jobs refactored
